requirements.txt
asn1crypto==0.23.0
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
cryptography==2.0.3
enum34==1.1.6
future==0.16.0
gevent==1.2.2
gevent-eventemitter==2.0
greenlet==0.4.12
idna==2.6
ipaddress==1.0.18
protobuf==3.4.0
pycparser==2.18
python-telegram-bot==8.0
requests==2.18.4
six==1.11.0
steam==0.8.21
urllib3==1.22
vdf==2.3
